+++ lib/slkecho/slack_request/post_message.rb
@@ -3,7 +3,7 @@
 module Slkecho
   module SlackRequest
     class PostMessage
-      Params = Struct.new(:channel, :message, :username, :icon_url, :icon_emoji, keyword_init: true)
+      Params = Struct.new(:channel, :blocks, :username, :icon_url, :icon_emoji, keyword_init: true)
 
       def initialize(slack_api_token:)
         @slack_api_token = slack_api_token
@@ -34,23 +34,11 @@ module Slkecho
       def request_body(params)
         {
           "channel" => params.channel,
-          "blocks" => blocks_from(params.message),
+          "blocks" => params.blocks,
           "username" => params.username,
           "icon_url" => params.icon_url,
           "icon_emoji" => params.icon_emoji
         }
-      end
-
-      def blocks_from(message)
-        [
-          {
-            "type" => "section",
-            "text" => {
-              "type" => "mrkdwn",
-              "text" => message
-            }
-          }
-        ]
       end
     end
   end